hhhh

+++ modular_skyrat/code/modules/surgery/bodyparts/_bodypart.dm
@@ -717,7 +717,7 @@
 	
 	for(var/datum/injury/IN in injuries)
 		// Wounds can disappear after 10 minutes at the earliest
-		if(IN.damage <= 0 && IN.created + (10 MINUTES) <= world.time)
+		if(IN.damage <= 0 && IN.created && IN.created + (10 MINUTES) <= world.time)
 			qdel(IN)
 			continue
